+++ tutors_assistant/admin/users/models/teacher.py
@@ -13,8 +13,8 @@ class Teacher(models.Model):
     telegram_name = models.CharField(
         verbose_name="Telegram name", max_length=200, blank=True
     )
-    telegram_id = models.IntegerField(verbose_name="Telegram id", blank=True)
-    telegram_chat = models.IntegerField(verbose_name="Telegram chat_id", blank=True)
+    telegram_id = models.BigIntegerField(verbose_name="Telegram id", blank=True)
+    telegram_chat = models.BigIntegerField(verbose_name="Telegram chat_id", blank=True)
     # Yandex
     yadisk_token = models.CharField(verbose_name="Токен яндекс диска", max_length=200, blank=True)
     yadisk_root_folder_name = models.CharField(verbose_name="Название корневой директори", max_length=200, blank=True)
